keypress Script
Allows a user to press keys 1, 2, 3, and 4 to switch from one webpage
to another. In order to use it:

1) Link to both the included jquery-3.1.1.min.js script and the
keyPressFunctions.js script in your HTML files. See the
website_horizontalRect pages for an example of how to do so.

2) Add any links from one page to another to the keyMap variable at the
top of the keyPressFunctions.js. (I will probably be the one to do
this, but if you want to add some yourself, follow the examples given.)
Currently, it can switch from index.html to register.html and back by
pressing the number 1.

Right now, all webpage mappings are included in the script file. This
can be changed to select values from a database later.