Horizontal Rectangle Version of the Website
This is just one possible version of the website, where the links are
placed in wide rectangular boxes and stacked on top of each other. It
includes only the index page and a super basic register HTML page for
now, along with the related CSS file. I will update this further later.